// generated/clients/javascript/services/uigf-game-cn-records/src/api/starRailApi/options.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

import { OperationOptions } from "@azure-rest/core-client";

/** Optional parameters. */
export interface StarRailApiGetChallengePeakOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface StarRailApiGetSimulatedUniverseOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface StarRailApiGetApocalypticShadowOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface StarRailApiGetPureFictionOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface StarRailApiGetForgottenHallOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface StarRailApiGetDailyNoteOptionalParams extends OperationOptions {}

/** Optional parameters. */
export interface StarRailApiGetIndexOptionalParams extends OperationOptions {}